+++ my-app/src/components/CardListItem/CardListItemSkeleton.stories.tsx
