+++ my-app/src/components/CardListItem/CardListItemSkeleton.tsx
