+++ my-app/src/components/CardListItem/EmptyCardIcon.tsx
